==> Scripts/2.AgregarNuevasViviendas.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd

# Cargar los archivos
df1 = pd.read_csv('/content/drive/MyDrive/Dataset Idealista/pisosBarcelona-04-04-2025.csv')
df2 = pd.read_csv('/content/drive/MyDrive/Dataset Idealista/pisos_barcelona.csv')

# Unir los dos DataFrames
df_combined = pd.concat([df1, df2], ignore_index=True)

# Eliminar duplicados basados en la columna 'propertycode'
df_cleaned = df_combined.drop_duplicates(subset='propertyCode')

# Guardar el resultado en un nuevo archivo CSV
output_path = '/content/drive/MyDrive/Dataset Idealista/pisos_barcelona_unificado.csv'
df_cleaned.to_csv(output_path, index=False)

print(f"Archivo guardado en: {output_path}")
